use decimal data type for population data
to ensure arithmetic in later models is accurate

+++ products/cdbg/models/staging/stg__low_mod_by_block_group.sql
@@ -8,7 +8,7 @@ SELECT
     "BORO" AS boro,
     "TRACT"::text AS tract,
     "BLKGRP"::text AS block_group,
-    REPLACE("LOWMODUNIV", ',', '')::integer AS potential_lowmod_population,
-    REPLACE("LOWMOD", ',', '')::integer AS low_mod_income_population,
+    REPLACE("LOWMODUNIV", ',', '')::decimal AS potential_lowmod_population,
+    REPLACE("LOWMOD", ',', '')::decimal AS low_mod_income_population,
     RTRIM("LOWMOD_PCT", '%')::numeric AS low_mod_income_population_percentage
 FROM {{ source("recipe_sources", "hud_lowmodincomebyblockgroup") }}
